Split out base class
